Point as an integer to speed this up.

+++ background.js
@@ -185,9 +185,9 @@ const schema = [
     name: 'points',
     type: SchemaTypes.array_ptr,
     schema: [
-      { name: 'x', type: SchemaTypes.f32 },
-      { name: 'y', type: SchemaTypes.f32 },
-      { name: 'tag', type: SchemaTypes.f32 },
+      { name: 'x', type: SchemaTypes.i32 },
+      { name: 'y', type: SchemaTypes.i32 },
+      { name: 'tag', type: SchemaTypes.i32 },
     ],
   },
 ];
